backend/requirements.txt
fastapi
uvicorn[standard]
pydantic
pydantic-settings
python-multipart
sqlalchemy
psycopg[binary]
python-jose[cryptography]
openai
